refactor: merge active and cancelled components into one
